// projects/library/feature-service/src/lib/tsc/enum/feature.enum.ts
export enum ApplicationFeature {
    NETZERO_ADMIN_FEATURE = 'NETZERO_ADMIN_FEATURE',
    ADMIN_FEATURE = 'ADMIN_FEATURE',
    NETZERO_FEATURE = 'NETZERO_FEATURE',
    ENERGY_FEATURE = 'ENERGY_FEATURE',
    WATER_FEATURE = 'WATER_FEATURE',
    ENVIRO_FEATURE = 'ENVIRO_FEATURE',
    CONNECT_FEATURE = 'CONNECT_FEATURE',
    OFFSET_FEATURE = 'OFFSET_FEATURE',
    SUPPLIER_FEATURE = 'SUPPLIER_FEATURE',
    EMPLOYEE_FEATURE = 'EMPLOYEE_FEATURE'
}